+++ tests/src/pbmer/test_Dbg.cpp
@@ -298,6 +298,74 @@ TEST(Pbmer_Dbg, FrequencyFilterNodes_empty)
     EXPECT_EQ(dg.NNodes(), 0);
 }
 
+TEST(Pbmer_Dbg, FrequencyFilterNodes2_empty)
+{
+
+    const PacBio::Pbmer::Parser parser{7};
+    //SNV                         (X)
+    const std::string td1{"ATGGAAGTCGCGGAACAAATC"};
+    const std::string td2{"ATGGAAGTGGCGGAACAAATC"};
+
+    PacBio::Pbmer::Mers m1{parser.Parse(td1)};
+    PacBio::Pbmer::Mers m2{parser.Parse(td2)};
+
+    PacBio::Pbmer::Dbg dg{7, 2};
+
+    dg.AddKmers(m1, 1);
+    dg.AddKmers(m2, 2);
+
+    dg.BuildEdges();
+
+    dg.FrequencyFilterNodes2(10);
+    EXPECT_EQ(dg.NNodes(), 0);
+}
+
+TEST(Pbmer_Dbg, FrequencyFilterNodes2_gt1)
+{
+
+    const PacBio::Pbmer::Parser parser{7};
+    //SNV                         (X)
+    const std::string td1{"ATGGAAGTCGCGGAACAAATC"};
+    const std::string td2{"ATGGAAGTGGCGGAACAAATC"};
+
+    PacBio::Pbmer::Mers m1{parser.Parse(td1)};
+    PacBio::Pbmer::Mers m2{parser.Parse(td2)};
+
+    PacBio::Pbmer::Dbg dg{7, 2};
+
+    dg.AddKmers(m1, 1);
+    dg.AddKmers(m2, 2);
+
+    dg.BuildEdges();
+
+    dg.FrequencyFilterNodes2(0, true);
+    EXPECT_EQ(dg.NNodes(), 0);
+}
+
+TEST(Pbmer_Dbg, FrequencyFilterNodes2_gt1_lt2)
+{
+
+    const PacBio::Pbmer::Parser parser{7};
+    //SNV                         (X)
+    const std::string td1{"ATGGAAGTCGCGGAACAAATC"};
+    const std::string td2{"ATGGAAGTGGCGGAACAAATC"};
+
+    PacBio::Pbmer::Mers m1{parser.Parse(td1)};
+    PacBio::Pbmer::Mers m2{parser.Parse(td2)};
+
+    PacBio::Pbmer::Dbg dg{7, 2};
+
+    dg.AddKmers(m1, 1);
+    dg.AddKmers(m2, 2);
+
+    dg.BuildEdges();
+
+    dg.FrequencyFilterNodes2(1, true);
+    dg.FrequencyFilterNodes2(1, false);
+    // 22 nodes 8 of which span the bubble path;
+    EXPECT_EQ(dg.NNodes(), 14);
+}
+
 TEST(Pbmer_Dbg, FrequencyFilterNodes_one_left)
 {
 
@@ -339,7 +407,7 @@ TEST(Pbmer_Dbg, GetNBreadth_bubble_found)
 
     dg.BuildEdges();
     auto bubbles = dg.GetBubbles();
-    EXPECT_EQ(bubbles.size(), 2);
+    EXPECT_EQ(bubbles.size(), 1);
 }
 
 TEST(Pbmer_Dbg, GetNBreadth_node_count_correct)
@@ -384,7 +452,7 @@ TEST(Pbmer_Dbg, GetNBreadth_bubble_found_rc_first)
     dg.BuildEdges();
     auto bubbles = dg.GetBubbles();
 
-    EXPECT_EQ(bubbles.size(), 2);
+    EXPECT_EQ(bubbles.size(), 1);
 }
 
 TEST(Pbmer_Dbg, GetNBreadth_bubble_found_rc)
@@ -406,7 +474,7 @@ TEST(Pbmer_Dbg, GetNBreadth_bubble_found_rc)
     dg.BuildEdges();
     auto bubbles = dg.GetBubbles();
 
-    EXPECT_EQ(bubbles.size(), 2);
+    EXPECT_EQ(bubbles.size(), 1);
 }
 
 TEST(Pbmer_Dbg, GetBubbles_mismatch_found)
@@ -427,7 +495,7 @@ TEST(Pbmer_Dbg, GetBubbles_mismatch_found)
     dg.BuildEdges();
     auto bubbles = dg.GetBubbles();
 
-    EXPECT_EQ(bubbles.size(), 2);
+    EXPECT_EQ(bubbles.size(), 1);
 }
 /*
 This is de bruijn graph has a complicated topology.Leave as a stretch
@@ -481,7 +549,7 @@ TEST(Pbmer_Dbg, GetBubbles_indel_found2)
     dg.BuildEdges();
     auto bubbles = dg.GetBubbles();
 
-    EXPECT_EQ(bubbles.size(), 2);
+    EXPECT_EQ(bubbles.size(), 1);
 }
 
 TEST(Pbmer_Dbg, GetBubbles_no_bubble)
@@ -623,7 +691,7 @@ TEST(Pbmer_Dbg, SpurRemoval_spur_bubble_cleanup)
     auto bubbles = dg.GetBubbles();
 
     EXPECT_EQ(spurCount, 1);
-    EXPECT_EQ(bubbles.size(), 2);
+    EXPECT_EQ(bubbles.size(), 1);
 }
 
 TEST(Pbmer_Dbg, GetBubbles_double_bubble_rc)
@@ -651,7 +719,7 @@ TEST(Pbmer_Dbg, GetBubbles_double_bubble_rc)
 
     auto bubbles = dg.GetBubbles();
 
-    EXPECT_EQ(bubbles.size(), 4);
+    EXPECT_EQ(bubbles.size(), 2);
 }
 
 TEST(Pbmer_Dbg, GetBubbles_complex_snv)
@@ -688,5 +756,5 @@ TEST(Pbmer_Dbg, GetBubbles_complex_snv)
 
     auto bubbles = dg.GetBubbles();
 
-    EXPECT_EQ(bubbles.size(), 2);
+    EXPECT_EQ(bubbles.size(), 1);
 }